requirements.txt
torch
torchvision
opencv-python
pandas
numpy
tqdm
scipy
pyyaml
matplotlib
scikit-learn
plotly
dash
pillow
